postOp gets dynamic context (#6)
* postOp gets dynamic context

postOp gets "bytes calldata context" instead of "bytes32" AND UserOperation.

rationale: in most cases, the entire UserOperation is not required by postOp, or at most, a single field is required.
OTOH, for complex paymasters, there might be some calculations in the payForOp that ARE required in the "postOp"

with the dynamic context, the paymaster decides exactly what to pass to its postOp, whether copy of field(s) form the UserOp, or locally calculated data.
If none is needed, it can be left empty.

- normal userOp cost ~400 more
- "create" operation cost 5000 less

* fixed test, initial geth testing support

+++ test/UserOperation.ts
@@ -6,10 +6,10 @@ export interface UserOperation {
   nonce: typ.uint256
   initCode: typ.bytes
   callData: typ.bytes
-  callGas: typ.uint64
+  callGas: typ.uint
   verificationGas: typ.uint
-  maxFeePerGas: typ.uint64
-  maxPriorityFeePerGas: typ.uint64
+  maxFeePerGas: typ.uint
+  maxPriorityFeePerGas: typ.uint
   paymaster: typ.address
   paymasterData: typ.bytes
   signer: typ.address
